token bucket patch

+++ src/or/hibernate.c
@@ -94,7 +94,7 @@ static uint64_t n_bytes_read_in_interval = 0;
 /** How many bytes have we written in this accounting interval? */
 static uint64_t n_bytes_written_in_interval = 0;
 /** How many seconds have we been running this interval? */
-static uint32_t n_seconds_active_in_interval = 0;
+static double n_seconds_active_in_interval = 0;
 /** How many seconds were we active in this interval before we hit our soft
  * limit? */
 static int n_seconds_to_hit_soft_limit = 0;
@@ -260,7 +260,7 @@ accounting_is_enabled(or_options_t *options)
  * passed, <b>n_read</b> bytes have been read, and <b>n_written</b>
  * bytes have been written. */
 void
-accounting_add_bytes(size_t n_read, size_t n_written, int seconds)
+accounting_add_bytes(size_t n_read, size_t n_written, double seconds)
 {
   n_bytes_read_in_interval += n_read;
   n_bytes_written_in_interval += n_written;
